[Refactor] EventDetailView의 Calendar 기능
- Data 를 IVO 패턴에 맞추어 이동하였습니다
